split Notification into client and server rendering
